Add PkiGadgetsCore API
