+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pvkgadgets",
-  "version": "0.0.8",
+  "version": "0.1.0",
   "private": true,
   "type": "module",
   "scripts": {
